Clear cargo clippy -D warnings across workspace (test lints + lib lints)
Automatic clippy --fix applied the mechanical test-code lints (useless vec!,
needless_borrow, double parens, etc.) across kernel/coin/settlement/atomic/node
test files, then manual fixes for what --fix leaves behind:

- x3-kernel: property_tests closure must-use Results bound via let _; removed
  unused MockDispatcher + collapsed identical if-blocks in mock; de-scaffolded
  unused alias and dead svm/evm_bytes in packet_integration_tests; replaced an
  always-true timestamp>=0 assertion with a real !=0 check.
- x3-coin: removed double parens; factored TeamVestingView tuple type alias.
- atomic-trade-engine: removed dead cross_vm_leg helper.
- x3-settlement-engine: unwrapped a redundant nested "mod tests"; tightened
  expired_locks type inference; kept the invariants multi-arg facade and the
  intrinsic-args dispatchable with documented clippy allows.
- runtime: dropped the deprecated RuntimeEvent assoc/cfg duplication, prefer
  iterators over index loops, derived Default for FreezeState, removed an
  identity op, and kept SDK-required runtime_version string/cfg idioms and the
  inline FraudProofV1 call payload with documented allows pending a scheduled
  runtime version/migration.
- x3-foundry-core unwrap and x3-kernel store Default impls landed earlier.

+++ pallets/x3-settlement-engine/tests/property_tests.rs
@@ -420,7 +420,7 @@ fn prop_zero_amount_settlement_valid() {
 
         // Should not panic or overflow
         prop_assert!(amount <= TOTAL_SUPPLY);
-        prop_assert!(legs_total <= MAX_SETTLEMENT_LEGS as u32);
+        prop_assert!(legs_total <= MAX_SETTLEMENT_LEGS);
     });
 }
